pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Alexander <alexsergeev45@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.3"
djangorestframework = "^3.13.1"
django-cors-headers = "^3.11.0"
psycopg2-binary = "^2.9.3"
django-environ = "^0.8.1"
Pillow = "^9.0.1"
PyJWT = "^2.3.0"
djangorestframework-simplejwt = "^5.1.0"
celery = "^5.2.5"
redis = "^4.2.1"
flower = "^1.0.0"
sentry-sdk = "^1.5.8"
django-simple-history = "^3.0.0"
django-filter = "^21.1"

[tool.poetry.dev-dependencies]
autopep8 = "^1.6.0"
flake8 = "^4.0.1"
Faker = "^13.3.4"
drf-yasg = "^1.20.0"
coverage = "^6.3.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
